requirements_d2rl_train.txt
async-timeout==4.0.3
attrs==23.1.0
click==8.0.4
cloudpickle==2.2.1
filelock==3.12.3
grpcio==1.43.0
pandas==1.4.3
gym==0.21.0
importlib-resources==6.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
msgpack==1.0.5
numpy==1.23.1
protobuf==3.20.3
PyYAML==6.0.1
dm-tree==0.1.8
ray[rllib]==1.11.0
redis==5.0.0
referencing==0.30.2
rpds-py==0.10.3
six==1.16.0
torch==1.11.0
tqdm==4.66.1
zipp==3.16.2
bidict==0.22.1
sparse==0.13.0
libsumo==1.18.0